requirements.txt
streamlit==1.29.0
Pillow==10.1.0
PyMuPDF==1.23.8
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
plotly==5.17.0